put gas heater controller in gas heater component
